WIP: query processor refactoring #63
