+++ CONTRIBUTING.md
@@ -5,14 +5,14 @@ Akumuli is free and open source project. Pull requests are always welcome.
 
 Please consider taking the following steps:
 
-1. Post to [Akumuli's mailing list](https://groups.google.com/forum/#!forum/akumuli) or gitter chat 
-before making any changes to keep efforts organized.
-2. Create issue with description of your enchacement or bug report (if your pull-request contains bug-fix). Mention this
+1. Create issue with description of your enchacement or bug report 
+(if your pull-request is a bug-fix). Mention this
 issue in your commit message if possible.
-3. __(important)__ Make changes to **dev** branch, not **master**.
+2. __(important)__ Make changes to **dev** branch, not **master**.
 
 Things to start with:
 
 1. Query processing: you can create new query processors relatively easily. 
-There are a lot of examples in `Akumuli/libakumuli/query_processing`. TBD: tutorial.
-2. New protocols. TBD: tutorial.
+There are a lot of examples in `Akumuli/libakumuli/query_processing`. 
+TBD: link to tutorial.
+2. New protocols. TBD: link to tutorial.
