+++ libakumuli/query_processing/randomsamplingnode.cpp
@@ -1,4 +1,5 @@
 #include "randomsamplingnode.h"
+#include <algorithm>
 
 namespace Akumuli {
 namespace QP {
