+++ libakumuli/seriesparser.cpp
@@ -89,251 +89,6 @@ void SeriesMatcher::pull_new_names(std::vector<SeriesMatcher::SeriesNameT> *buff
     std::swap(names, *buffer);
 }
 
-static boost::optional<std::string> parse_select_stmt(boost::property_tree::ptree const& ptree, aku_logger_cb_t logger) {
-    auto select = ptree.get_child_optional("select");
-    if (select && select->empty()) {
-        // simple select query
-        auto str = select->get_value<std::string>("");
-        if (str == "names") {
-            // the only supported select query for now
-            return str;
-        }
-        (*logger)(AKU_LOG_ERROR, "Invalid `select` query");
-        auto rte = std::runtime_error("Invalid `select` query");
-        BOOST_THROW_EXCEPTION(rte);
-    }
-    return boost::optional<std::string>();
-}
-
-static QP::GroupByStatement parse_groupby(boost::property_tree::ptree const& ptree,
-                                          aku_logger_cb_t logger) {
-    aku_Timestamp duration = 0u;
-    auto groupby = ptree.get_child_optional("group-by");
-    if (groupby) {
-        for(auto child: *groupby) {
-            if (child.first == "time") {
-                std::string str = child.second.get_value<std::string>();
-                duration = DateTimeUtil::parse_duration(str.c_str(), str.size());
-            }
-        }
-    }
-    return QP::GroupByStatement(duration);
-}
-
-static std::vector<std::string> parse_metric(boost::property_tree::ptree const& ptree,
-                                             aku_logger_cb_t logger) {
-    std::vector<std::string> metrics;
-    auto metric = ptree.get_child_optional("metric");
-    if (metric) {
-        auto single = metric->get_value<std::string>();
-        if (single.empty()) {
-            for(auto child: *metric) {
-                auto metric_name = child.second.get_value<std::string>();
-                metrics.push_back(metric_name);
-            }
-        } else {
-            metrics.push_back(single);
-        }
-    }
-    return metrics;
-}
-
-static aku_Timestamp parse_range_timestamp(boost::property_tree::ptree const& ptree,
-                                           std::string const& name,
-                                           aku_logger_cb_t logger) {
-    auto range = ptree.get_child("range");
-    for(auto child: range) {
-        if (child.first == name) {
-            auto iso_string = child.second.get_value<std::string>();
-            auto ts = DateTimeUtil::from_iso_string(iso_string.c_str());
-            return ts;
-        }
-    }
-    std::stringstream fmt;
-    fmt << "can't find `" << name << "` tag inside the query";
-    QueryParserError error(fmt.str().c_str());
-    BOOST_THROW_EXCEPTION(error);
-}
-
-
-static aku_ParamId extract_id_from_pool(StringPool::StringT res) {
-    // Series name in string pool should be followed by \0 character and 64-bit series id.
-    auto p = res.first + res.second;
-    assert(p[0] == '\0');
-    p += 1;  // zero terminator + sizeof(uint64_t)
-    return *reinterpret_cast<uint64_t const*>(p);
-}
-
-static std::vector<aku_ParamId> parse_where_clause(boost::property_tree::ptree const& ptree,
-                                                   std::string metric,
-                                                   std::string pred,
-                                                   StringPool const& pool,
-                                                   aku_logger_cb_t logger)
-{
-    std::vector<aku_ParamId> ids;
-    bool not_set = false;
-    auto where = ptree.get_child_optional("where");
-    if (where) {
-        for (auto child: *where) {
-            auto predicate = child.second;
-            auto items = predicate.get_child_optional(pred);
-            if (items) {
-                for (auto item: *items) {
-                    std::string tag = item.first;
-                    auto idslist = item.second;
-                    // Read idlist
-                    for (auto idnode: idslist) {
-                        std::string value = idnode.second.get_value<std::string>();
-                        std::stringstream series_regexp;
-                        series_regexp << "(" << metric << R"((?:\s\w+=\w+)*\s)"
-                                      << tag << "=" << value << R"((?:\s\w+=\w+)*))";
-                        std::string regex = series_regexp.str();
-                        auto results = pool.regex_match(regex.c_str());
-                        for(auto res: results) {
-                            aku_ParamId id = extract_id_from_pool(res);
-                            ids.push_back(id);
-                        }
-                    }
-                }
-            } else {
-                not_set = true;
-            }
-        }
-    } else {
-        not_set = true;
-    }
-    if (not_set) {
-        if (pred == "in") {
-            // there is no "in" predicate so we need to include all
-            // series from this metric
-            std::stringstream series_regexp;
-            series_regexp << "" << metric << R"((\s\w+=\w+)*)";
-            std::string regex = series_regexp.str();
-            auto results = pool.regex_match(regex.c_str());
-            for(auto res: results) {
-                aku_ParamId id = extract_id_from_pool(res);
-                ids.push_back(id);
-            }
-        }
-    }
-    return ids;
-}
-
-static std::string to_json(boost::property_tree::ptree const& ptree, bool pretty_print = true) {
-    std::stringstream ss;
-    boost::property_tree::write_json(ss, ptree, pretty_print);
-    return ss.str();
-}
-
-std::shared_ptr<QP::IQueryProcessor>
-SeriesMatcher::build_query_processor(const char* query, std::shared_ptr<QP::Node> terminal, aku_logger_cb_t logger) {
-    namespace pt = boost::property_tree;
-    using namespace QP;
-
-    const auto NOSAMPLE = std::make_pair<std::string, size_t>("", 0u);
-
-    //! C-string to streambuf adapter
-    struct MemStreambuf : std::streambuf {
-        MemStreambuf(const char* buf) {
-            char* p = const_cast<char*>(buf);
-            setg(p, p, p+strlen(p));
-        }
-    };
-
-    boost::property_tree::ptree ptree;
-    MemStreambuf strbuf(query);
-    std::istream stream(&strbuf);
-    try {
-        pt::json_parser::read_json(stream, ptree);
-    } catch (pt::json_parser_error& e) {
-        // Error, bad query
-        (*logger)(AKU_LOG_ERROR, e.what());
-        throw QueryParserError(e.what());
-    }
-
-    logger(AKU_LOG_INFO, "Parsing query:");
-    logger(AKU_LOG_INFO, to_json(ptree, true).c_str());
-
-    try {
-        // Read groupby statement
-        auto groupby = parse_groupby(ptree, logger);
-
-        // Read metric(s) name
-        auto metrics = parse_metric(ptree, logger);
-
-        // Read select statment
-        auto select = parse_select_stmt(ptree, logger);
-
-        // Read sampling method
-        auto sampling_params = ptree.get_child_optional("sample");
-
-        // Read where clause
-        std::vector<aku_ParamId> ids_included;
-        std::vector<aku_ParamId> ids_excluded;
-
-        for(auto metric: metrics) {
-
-            auto in = parse_where_clause(ptree, metric, "in", pool, logger);
-            std::copy(in.begin(), in.end(), std::back_inserter(ids_included));
-
-            auto notin = parse_where_clause(ptree, metric, "not_in", pool, logger);
-            std::copy(notin.begin(), notin.end(), std::back_inserter(ids_excluded));
-        }
-
-        if (sampling_params && select) {
-            (*logger)(AKU_LOG_ERROR, "Can't combine select and sample statements together");
-            auto rte = std::runtime_error("`sample` and `select` can't be used together");
-            BOOST_THROW_EXCEPTION(rte);
-        }
-
-        // Build topology
-        std::shared_ptr<Node> next = terminal;
-        if (!select) {
-            // Read timestamps
-            auto ts_begin = parse_range_timestamp(ptree, "from", logger);
-            auto ts_end = parse_range_timestamp(ptree, "to", logger);
-
-            if (sampling_params) {
-                for (auto i = sampling_params->rbegin(); i != sampling_params->rend(); i++) {
-                        next = NodeBuilder::make_sampler(i->second,
-                                                         next,
-                                                         logger);
-                }
-            }
-            if (!ids_included.empty()) {
-                next = NodeBuilder::make_filter_by_id_list(ids_included, next, logger);
-            }
-            if (!ids_excluded.empty()) {
-                next = NodeBuilder::make_filter_out_by_id_list(ids_excluded, next, logger);
-            }
-            // Build query processor
-            return std::make_shared<ScanQueryProcessor>(next, metrics, ts_begin, ts_end, groupby);
-        }
-
-        if (ids_included.empty() && metrics.empty()) {
-            // list all
-            for (auto val: table) {
-                auto id = val.second;
-                ids_included.push_back(id);
-            }
-        }
-        if (!ids_excluded.empty()) {
-            std::sort(ids_included.begin(), ids_included.end());
-            std::sort(ids_excluded.begin(), ids_excluded.end());
-            std::vector<aku_ParamId> tmp;
-            std::set_difference(ids_included.begin(), ids_included.end(),
-                                ids_excluded.begin(), ids_excluded.end(),
-                                std::back_inserter(tmp));
-            std::swap(tmp, ids_included);
-        }
-        return std::make_shared<MetadataQueryProcessor>(ids_included, next);
-
-    } catch(std::exception const& e) {
-        (*logger)(AKU_LOG_ERROR, e.what());
-        throw QueryParserError(e.what());
-    }
-}
-
 //                         //
 //      Series Parser      //
 //                         //
